Fix jobs resilience to db replication delay
Summary: We postpone the job for 5 seconds if the record couldn't be found

Test Plan: `./phpunit`

Reviewers: #hkccp_developers, vanmeeuwen

Reviewed By: #hkccp_developers, vanmeeuwen

Subscribers: #hkccp_developers

Differential Revision: https://git.kolab.org/D2359

+++ src/app/Jobs/Group/UpdateJob.php
@@ -15,6 +15,10 @@ class UpdateJob extends GroupJob
     {
         $group = $this->getGroup();
 
+        if (!$group) {
+            return;
+        }
+
         if (!$group->isLdapReady()) {
             $this->delete();
             return;
